refactored: packages for better structure
