+++ src/main/java/de/nikoconsulting/demo/multitenancy/repositories/GenericEntityRepository.java
@@ -1,5 +1,6 @@
-package de.nikoconsulting.demo.multitenancy;
+package de.nikoconsulting.demo.multitenancy.repositories;
 
+import de.nikoconsulting.demo.multitenancy.model.GenericEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface GenericEntityRepository extends JpaRepository<GenericEntity, Long> {
